move methods shared by TorService, and TestTorService to BaseService
