+++ topl-service/src/main/java/io/matthewnelson/topl_service/service/components/binding/TorServiceConnection.kt
@@ -72,8 +72,7 @@ import android.os.IBinder
 internal class TorServiceConnection: BaseServiceConnection() {
 
     companion object {
-        val torServiceConnection =
-            TorServiceConnection()
+        val torServiceConnection = TorServiceConnection()
     }
 
     override fun onServiceDisconnected(name: ComponentName?) {
